chore: Move handing of the Advanced Search whitelist to the SearchProxy, drop ENV-based configuration
