+++ models/classes/AdvancedSearch/AdvancedSearchChecker.php
@@ -27,43 +27,16 @@ use oat\tao\model\featureFlag\FeatureFlagChecker;
 use oat\tao\model\featureFlag\FeatureFlagCheckerInterface;
 use oat\tao\model\search\SearchInterface;
 use oat\tao\model\search\SearchProxy;
-use common_ext_ExtensionsManager;
 
 class AdvancedSearchChecker extends ConfigurableService
 {
     /** @deprecated Use oat\tao\model\featureFlag\FeatureFlagCheckerInterface::FEATURE_FLAG_ADVANCED_SEARCH_DISABLED */
     public const FEATURE_FLAG_ADVANCED_SEARCH_DISABLED = FeatureFlagCheckerInterface::FEATURE_FLAG_ADVANCED_SEARCH_DISABLED;
 
-    public const CONFIG_ADVANCED_SEARCH_DISABLED_SECTIONS = 'TAO_ADVANCED_SEARCH_DISABLED_SECTIONS';
-
     public function isEnabled(): bool
     {
         return !$this->getFeatureFlagChecker()->isEnabled(FeatureFlagCheckerInterface::FEATURE_FLAG_ADVANCED_SEARCH_DISABLED)
             && $this->getSearchService()->supportCustomIndex();
-    }
-
-    public function getDisabledSections(): array
-    {
-        $disabledSections = [
-            'results',
-        ];
-
-        if ($this->getExtensionsManager()->isEnabled('taoBooklet')) {
-            $disabledSections[] = 'taoBooklet_main';
-        }
-
-        if (isset($_ENV[self::CONFIG_ADVANCED_SEARCH_DISABLED_SECTIONS])) {
-            $conf = $_ENV[self::CONFIG_ADVANCED_SEARCH_DISABLED_SECTIONS];
-
-            foreach (explode(',', $conf) as $section) {
-                $section = trim($section);
-                if ($section != '' && !in_array($section, $disabledSections)) {
-                    $disabledSections[] = $section;
-                }
-            }
-        }
-
-        return $disabledSections;
     }
 
     private function getFeatureFlagChecker(): FeatureFlagCheckerInterface
@@ -74,10 +47,5 @@ class AdvancedSearchChecker extends ConfigurableService
     private function getSearchService(): SearchInterface
     {
         return $this->getServiceLocator()->get(SearchProxy::SERVICE_ID);
-    }
-
-    private function getExtensionsManager(): common_ext_ExtensionsManager
-    {
-        return $this->getServiceLocator()->get(common_ext_ExtensionsManager::SERVICE_ID);
     }
 }
